moved gpt handler to loading fragment from llm response fragment
